+++ ft_printf_utils_ptr.c
@@ -12,7 +12,7 @@
 
 #include "ft_printf.h"
 
-static int	ft_ptr_len(uintptr_t num)
+static int	ft_ptr_len(unsigned long num)
 {
 	int	len;
 
@@ -25,7 +25,7 @@ static int	ft_ptr_len(uintptr_t num)
 	return (len);
 }
 
-static void	ft_put_ptr(uintptr_t num)
+static void	ft_put_ptr(unsigned long num)
 {
 	if (num >= 16)
 	{
@@ -41,7 +41,7 @@ static void	ft_put_ptr(uintptr_t num)
 	}
 }
 
-int	ft_print_ptr(unsigned long long ptr)
+int	ft_print_ptr(unsigned long ptr)
 {
 	int	print_length;
